restructured important files
