+++ src/perception/lidar_pipeline_3/lidar_pipeline_3/library/utils.py
@@ -5,7 +5,7 @@ from logging import Logger  # For typing
 import os
 import shutil
 
-from . import constants as const
+from .. import constants as const
 
 from typing import Any, Tuple
