+++ src/perception/lidar_pipeline_3/lidar_pipeline_3/library/video_stitcher.py
