+++ src/perception/lidar_pipeline_3/lidar_pipeline_3/library/visualiser.py
@@ -5,7 +5,6 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from .. import constants as const
-from .. import utils
 
 
 def get_segment_count(segments):
